book-command-service/requirements.txt
flask>=3.0
mysql-connector-python>=9.0
google-generativeai>=0.8
